requirements.txt
blinker
certifi
cffi
charset-normalizer
click
cryptography
Flask
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
pyTelegramBotAPI
python-dotenv
requests
schedule
SQLAlchemy
telebot
typing_extensions
urllib3
Werkzeug

